+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Started with a benchmark module
 - Introduced byte code instructions for "." and "~" for better performance, and more readable byte code.
 - Introduced byte code instructions "until_char" and "until_set" to speed up "find", "keepto" and "findall" macros
+- Compiler supports manually overriding which variables are captured. 
 
 ## V 0.1.9
 - Fixed issues with "(a)+" like pattern. See https://gitlab.com/rosie-pattern-language/rosie/-/issues/123